+++ funniture-fe/src/apis/AdminInquiryAPI.js
@@ -22,6 +22,14 @@ export async function getUSerInquiryList({ memberId }) {
     }
 }
 
+export async function sendChat({ newChat }) {
+    console.log("API 전달 전 : ", newChat)
+
+    const url = `/adinquiry/sendChat`
+
+    await api.post(url, newChat)
+}
+
 // 공용
 const getData = async (url, query) => {
     let response
