+++ funniture-fe/src/component/ChatModal.module.css
@@ -219,14 +219,22 @@
         overflow: scroll;
 
         .directChatItem {
-            max-width: 70%;
+            max-width: 80%;
             /* min-height: 40px; */
             display: flex;
             align-items: end;
             margin: 5px 0;
+            font-size: 0.8em;
+            justify-content: end;
 
-            >div {
+            /* >div {
                 height: 100%;
+            } */
+
+            .sendTime {
+                text-align: end;
+                max-width: 20%;
+                font-size: 0.5em;
             }
 
             .chatContent {
@@ -280,7 +288,7 @@
     padding-top: 0;
 
     min-height: 12%;
-    max-height: 15%;
+    max-height: 18%;
 
     .inputBox {
         width: 100%;
