+++ funniture-fe/src/pages/admin/AdminDirectPage.module.css
@@ -77,7 +77,7 @@
         .orderByBox {
             width: 100%;
             display: flex;
-            justify-content: end;
+            justify-content: space-between;
             border-top: solid 1px #7F5539;
             border-bottom: solid 1px #7F5539;
             padding: 8px 5px;
@@ -87,6 +87,23 @@
                 border: none;
                 background-color: #F6EEE8;
                 font-size: 0.8em;
+            }
+
+            >div {
+                display: flex;
+
+                button {
+                    border: none;
+                    background: none;
+                    font-size: 0.8em;
+                    color: #797979;
+
+                    margin-right: 5px;
+                }
+
+                img {
+                    width: 0.8em;
+                }
             }
         }
